Don't Add Management Network to Edge data_network_ids
This is because data networks are for overlay and vlan uplink. The
management network is for NSX MP to Edge connection. The management
network should not be used as data network (in theory). If in practice
the management network and vlan uplink network are the same, the user
still can specify the same value for vc_uplink_network_for_edge
and vc_management_network_for_edge.

Signed-off-by: Zhengsheng Zhou <zhengshengz@vmware.com>